backend/requirements.txt
python-chess==0.15.3
flask==0.11.1
simplejson
